[ci] fix pytest ci action failure
Pytest returns exit code 5 whenever it doesn't detect any tests, or skip
them all. In theory, this was done to let people know about broken CI
setups. When `tests/test_basic.py` was renamed to `tests/test_basic.py.old`
(likely due to `new_defm` not being exposed/extant in the epiworldpy
Python module), pytest started failing to detect any tests, failing the
CI action. To fix this we remove the code that calls `new_defm` (which
shouldn't be an issue, since the it was .old'd), and only keep the
version check.