[ghci] remove test cruft
Previously, there were a lot of superfluous tests. They were building
wheels and testing packages multiple times, as well as building WASM
binaries. Most of these weren't even run, because they would only be
dispatched on a change to the nonexistant "master" branch (this
repository uses "main").

The only functionality that might be missed is the ability to create a
release automatically, but that seems sufficiently far off in the future
to not worry about that for now.

The tests that have been kept now test against a lesser array of Python
versions (3.7 and 3.11 in this case), as well as only testing on Ubuntu
instaid of Ubuntu, Windows, and macOS.

Fixes #12.