[bind] optimize, expose those mentioned in #20
Here, we expose the functions mentioned in issue #20. We also do some
optimization; namely, before, lots of data was being copied over to
Python. Now, we bind the data to a Python capsule and then convert that
over to an array. The underlying C++ vector will only be destroyed when
the Python object sharing the data goes out of scope. This doesn't seem
to be documented within PyBind, but from my own testing it appears to
work just fine.

It's worth noting that I couldn't find a way to do this while
maintaining some semblance of API similarity with the function
`get_reproductive_number`, due to the fact that the values it returns
are not "rectangular" and thus cannot be represented by Python arrays
without having to insert some filler values.