[bind] expose `run_multiple` and `Saver`
See #23. There's a bunch of filesystem stuff here, since we're on C++11
and don't have access to the C++17 filesytem APIs. The only thing we
should have to worry about here is that the implementation of the Saver
constructor matches the Saver constructor (or equivilant) in EpiworldR:

https://github.com/UofUEpiBio/epiworldR/blob/main/R/make_saver.R

The thing that most needs to be watched out for is what happens if you
supply `fn` a file versus a directory; I've done some testing and it
appears to be working fine, but there might be some edge cases I didn't
test. I would write a unit test to check, but I'm not entirely sure
about the intended behaviour of the original R code.