[bandit/codacy] ignore use of `assert`
This makes Bandit (and by extension Codacy) ignore the use of `assert`.
Codacy can't seem to tell test from prod, so whenever `assert` is used
in the test cases it will fail CI. This fix is not ideal (we'd still
like to be warned about `assert` in real non-test code), but there
appears to be no other way to fix it outside of mucking about in the
admin portal (which would achieve no better effect).

Reference:
https://github.com/fossasia/query-server/issues/332